Identify Non-User Downloading
When a user not registered with a site is shared a link, the download of non registered users should be identified in the activity history in an identifiable way.  Anonymous users should have their IP's logged in history so that administrators can determine if the content is being downloaded by the originally intended party or if the information was forward beyond the original intention.  It also provides a method to get an idea which users have downloaded the file (intended) and when.